data-services/ingestion/unzip/requirements.txt
functions-framework==3.*
pandas
google-cloud-storage
pyarrow
fsspec
gcsfs